Implement Pagination component for Students page
